+++ src/entrypoints/content/enhancers/pt/index.ts
@@ -1,0 +1,5 @@
+/**
+ * PT 模块单一出口
+ */
+
+export { PTDimmer } from './dimmer'